requirements.txt
Django>=5.2.0
djangorestframework>=3.16.0
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0
joblib>=1.0.0
openpyxl>=3.0.0
pypfopt>=1.5.0
yfinance>=0.2.0